// core/utils/ui_utils.d.ts
export declare const Colors: {
    Reset: string;
    Bright: string;
    Dim: string;
    Underscore: string;
    Blink: string;
    Reverse: string;
    Hidden: string;
    FgBlack: string;
    FgRed: string;
    FgGreen: string;
    FgYellow: string;
    FgBlue: string;
    FgMagenta: string;
    FgCyan: string;
    FgWhite: string;
    BgBlack: string;
    BgRed: string;
    BgGreen: string;
    BgYellow: string;
    BgBlue: string;
    BgMagenta: string;
    BgCyan: string;
    BgWhite: string;
};
export declare function colorize(text: string, color: string): string;
export declare function displayRitualStepResult(res: any): void;
export declare function demonstrateCursorControl(): void;
export declare function startCursorAnimation(): void;
export declare function stopCursorAnimation(): void;